chore: 파일 변경사항

+++ src/main/java/com/sprint/hrbank/application/department/provided/command/DepartmentCleaner.java
@@ -1,5 +1,5 @@
 package com.sprint.hrbank.application.department.provided.command;
 
 public interface DepartmentCleaner {
-  void delete(Integer departmentId);
+  void delete(Long departmentId);
 }
