+++ src/main/java/com/sprint/hrbank/application/department/provided/command/DepartmentModifier.java
@@ -4,5 +4,5 @@ import com.sprint.hrbank.application.department.dto.DepartmentDto;
 import com.sprint.hrbank.application.department.dto.DepartmentUpdateRequest;
 
 public interface DepartmentModifier {
-  DepartmentDto update(Integer departmentId, DepartmentUpdateRequest updateRequest);
+  DepartmentDto update(Long departmentId, DepartmentUpdateRequest updateRequest);
 }
