+++ src/main/java/com/sprint/hrbank/application/department/provided/query/DepartmentFinder.java
@@ -4,5 +4,5 @@ import com.sprint.hrbank.domain.department.Department;
 
 public interface DepartmentFinder {
 
-  Department getById(Integer departmentId);
+  Department getById(Long departmentId);
 }
